requirements.txt
# Core dependencies for DAViD models
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0  # For GPU support
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0 